pyproject.toml
[project]
name = "hr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.7.0",
    "numpy>=2.2.4",
    "scripts>=3.0",
    "stopwatch-py>=2.0.1",
]
